fix: Update IoxMetadata to shard_id; supporting transition
This is the custom metadata stored in parquet files. Rather than
reserving the field in the proto (which removes it from the Rust
structs), change it to optional and support reading either field.

If we decide to delete all parquet files that use sequencer_id, we can
remove it and only use shard_id at that point.

Connects to #4246.